refactor (remove package) : Interface 패키지 지움
