+++ src/main/java/kitchenpos/bo/TableBo.java
@@ -1,7 +1,7 @@
 package kitchenpos.bo;
 
-import kitchenpos.dao.Interface.OrderDao;
-import kitchenpos.dao.Interface.OrderTableDao;
+import kitchenpos.dao.OrderDao;
+import kitchenpos.dao.OrderTableDao;
 import kitchenpos.model.OrderStatus;
 import kitchenpos.model.OrderTable;
 import org.springframework.stereotype.Component;
