+++ src/main/java/kitchenpos/dao/DefaultMenuProductDao.java
@@ -1,6 +1,5 @@
 package kitchenpos.dao;
 
-import kitchenpos.dao.Interface.MenuProductDao;
 import kitchenpos.model.MenuProduct;
 import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.jdbc.core.namedparam.BeanPropertySqlParameterSource;
