+++ src/main/java/kitchenpos/dao/DefaultOrderDao.java
@@ -1,6 +1,5 @@
 package kitchenpos.dao;
 
-import kitchenpos.dao.Interface.OrderDao;
 import kitchenpos.model.Order;
 import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.jdbc.core.namedparam.BeanPropertySqlParameterSource;
