+++ src/main/java/kitchenpos/dao/DefaultOrderLineItemDao.java
@@ -16,7 +16,7 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public class DefaultOrderLineItemDao implements kitchenpos.dao.Interface.OrderLineItemDao {
+public class DefaultOrderLineItemDao implements OrderLineItemDao {
     private static final String TABLE_NAME = "order_line_item";
     private static final String KEY_COLUMN_NAME = "seq";
